Investigated issues with mclk pin
From now, setting the mclk pin by user is disabled because
this ability creates more problem than it solves.
Also, I've performed an investigation on this topic
and found out that there's OC1A output attached to
digital pin 9 of Arduino Uno Rev3 so it's the only
way to access internal clocking system.